added abstract classes to define what model/controllers should look like.
