+++ seamCarvingJava/src/controller/CustomController.java
@@ -1,0 +1,17 @@
+package controller;
+
+import javafx.stage.Stage;
+
+public abstract class CustomController {
+
+	protected Stage _primaryStage;
+	
+	public void init(Stage primaryStage) {
+		_primaryStage = primaryStage;
+		_primaryStage.show();
+		start();
+	}
+	
+	protected abstract void start();
+	
+}
