+++ seamCarvingJava/src/model/CustomModel.java
@@ -1,0 +1,9 @@
+package model;
+
+import javafx.stage.Stage;
+
+public abstract class CustomModel {
+	
+	public abstract void start(Stage primaryStage);
+	
+}
